theseus_runner/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pygame>=2.5.0
